backend-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pydantic-settings==2.7.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
